requirements.txt
Django==2.0.6
Pillow==5.1.0
djangorestframework==3.8.2
python-decouple==3.1